添加额外破坏效果配置

+++ src/main/java/com/atsuishio/superbwarfare/entity/projectile/C4Entity.java
@@ -419,11 +419,11 @@ public class C4Entity extends Entity implements GeoEntity, OwnableEntity {
             }
         }
 
-        if (this.level() instanceof ServerLevel) {
+        if (this.level() instanceof ServerLevel && ExplosionConfig.EXPLOSION_DESTROY.get() && ExplosionConfig.EXTRA_EXPLOSION_EFFECT.get()) {
             AABB aabb = new AABB(pos, pos).inflate(2);
             BlockPos.betweenClosedStream(aabb).forEach((blockPos) -> {
                 float hard = this.level().getBlockState(blockPos).getBlock().defaultDestroyTime();
-                if (ExplosionConfig.EXPLOSION_DESTROY.get() && hard != -1) {
+                if (hard != -1) {
                     this.level().destroyBlock(blockPos, true);
                 }
             });
